+++ data/src/client.rs
@@ -1,5 +1,5 @@
 use std::cmp::Ordering;
-use std::collections::{BTreeMap, HashMap, HashSet, VecDeque};
+use std::collections::{BTreeMap, HashMap, HashSet};
 use std::mem::take;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -26,10 +26,10 @@ use crate::environment::{SOURCE_WEBSITE, VERSION};
 use crate::features::{self, Features, VersionRequest};
 use crate::history::ReadMarker;
 use crate::isupport::{
-    ChatHistoryState, ChatHistorySubcommand, MessageReference, WhoToken,
-    WhoXPollParameters, find_target_limit,
+    ChatHistoryState, ChatHistorySubcommand, MessageReference,
+    find_target_limit,
 };
-use crate::rate_limit::{BackoffInterval, TokenBucket, TokenPriority};
+use crate::rate_limit::{TokenBucket, TokenPriority};
 use crate::target::{self, Target};
 use crate::time::Posix;
 use crate::user::{ChannelUsers, Nick, NickRef};
@@ -40,6 +40,9 @@ use crate::{
 };
 
 pub mod on_connect;
+pub mod who_queue;
+
+use who_queue::{WhoSource, WhoStatus};
 
 const HIGHLIGHT_BLACKOUT_INTERVAL: Duration = Duration::from_secs(5);
 const CLIENT_CHATHISTORY_LIMIT: u16 = 500;
@@ -207,8 +210,7 @@ pub struct Client {
     notification_blackout: NotificationBlackout,
     registration_required_channels: Vec<target::Channel>,
     isupport: HashMap<isupport::Kind, isupport::Parameter>,
-    who_polls: VecDeque<WhoPoll>,
-    who_poll_interval: BackoffInterval,
+    who_queue: who_queue::WhoQueue,
     resolved_netid: Option<String>,
     anti_flood: Option<TokenBucket<message::Encoded>>,
     mode_requests: Vec<ModeRequest>,
@@ -272,12 +274,7 @@ impl Client {
             ),
             registration_required_channels: vec![],
             isupport: HashMap::new(),
-            who_polls: VecDeque::new(),
-            who_poll_interval: BackoffInterval::from(
-                config
-                    .who_poll_interval
-                    .min(config.anti_flood.saturating_mul(2)),
-            ),
+            who_queue: who_queue::WhoQueue::new(config.clone()),
             resolved_netid: None,
             anti_flood: Some(TokenBucket::new(config.anti_flood, 10)),
             mode_requests: Vec::new(),
@@ -448,25 +445,6 @@ impl Client {
             return events;
         }
 
-        if self.config.who_poll_enabled != config.who_poll_enabled {
-            if config.who_poll_enabled {
-                for channel in self.chanmap.keys() {
-                    self.who_polls.push_back(WhoPoll {
-                        channel: channel.clone(),
-                        status: WhoStatus::Joined,
-                    });
-                }
-            } else {
-                self.who_polls.retain(|who_poll| {
-                    matches!(
-                        who_poll.status,
-                        WhoStatus::Requested(_, _, _)
-                            | WhoStatus::Receiving(_, _)
-                    )
-                });
-            }
-        }
-
         if self.config.order_channels_by != config.order_channels_by
             || self.config.channels != config.channels
         {
@@ -483,18 +461,15 @@ impl Client {
             });
         }
 
+        self.who_queue.update_config(&self.chanmap, config.clone());
+
         self.config = config;
 
         events
     }
 
     fn quit(&mut self, reason: Option<String>) {
-        self.who_polls.retain(|who_poll| {
-            matches!(
-                who_poll.status,
-                WhoStatus::Requested(_, _, _) | WhoStatus::Receiving(_, _)
-            )
-        });
+        self.who_queue.quit();
 
         if let Err(e) = if let Some(reason) = reason {
             self.handle.try_send(command!("QUIT", reason))
@@ -555,7 +530,7 @@ impl Client {
                     // should be in who_polls and rerouting should be stopped, or it
                     // is treated as part of a rate-limiting response and if there
                     // are any outstanding user requests rerouting should be stopped.
-                    self.who_polls.iter().any(|who_poll| {
+                    self.who_queue.any_matching(|who_poll| {
                         matches!(
                             who_poll.status,
                             WhoStatus::Requested(WhoSource::User, _, _)
@@ -665,25 +640,8 @@ impl Client {
                         };
 
                         if let Some(channel) = channel {
-                            // Record user WHO request(s) for reply filtering
-                            let status = WhoStatus::Requested(
-                                WhoSource::User,
-                                Instant::now(),
-                                params.get(2).and_then(|token| {
-                                    token.parse::<WhoToken>().ok()
-                                }),
-                            );
-
-                            if let Some(who_poll) = self
-                                .who_polls
-                                .iter_mut()
-                                .find(|who_poll| who_poll.channel == channel)
-                            {
-                                who_poll.status = status;
-                            } else {
-                                self.who_polls
-                                    .push_front(WhoPoll { channel, status });
-                            }
+                            self.who_queue
+                                .requested_channel_poll(params, channel);
                         }
                     }
                 }
@@ -1741,7 +1699,8 @@ impl Client {
                 })]);
             }
             Command::PART(channel, _) => {
-                let user = ok!(message.user(self.casemapping()));
+                let casemapping = self.casemapping();
+                let user = ok!(message.user(casemapping));
 
                 if user.nickname() == self.nickname() {
                     self.chanmap.shift_remove(&context!(
@@ -1752,6 +1711,23 @@ impl Client {
                             self.casemapping(),
                         )
                     ));
+
+                    // Remove departed channel who polls from queue
+                    let target_channel = context!(target::Channel::parse(
+                        channel,
+                        self.chantypes(),
+                        self.statusmsg(),
+                        casemapping,
+                    ));
+                    self.who_queue.remove_matching(|who_poll| {
+                        who_poll.channel == target_channel
+                            && (matches!(
+                                who_poll.status,
+                                WhoStatus::Joined
+                                    | WhoStatus::Received
+                                    | WhoStatus::Waiting(_)
+                            ))
+                    });
                 } else if let Some(channel) =
                     self.chanmap.get_mut(&context!(target::Channel::parse(
                         channel,
@@ -1793,16 +1769,7 @@ impl Client {
                     );
 
                     // Add channel to WHO poll queue
-                    if !self
-                        .who_polls
-                        .iter()
-                        .any(|who_poll| who_poll.channel == target_channel)
-                    {
-                        self.who_polls.push_back(WhoPoll {
-                            channel: target_channel.clone(),
-                            status: WhoStatus::Joined,
-                        });
-                    }
+                    self.who_queue.queue_channel_poll(&target_channel);
 
                     if !self.mode_requests.iter().any(|mode_request| {
                         mode_request.channel == target_channel
@@ -1888,34 +1855,50 @@ impl Client {
                     self.statusmsg(),
                     casemapping,
                 ) {
-                    let user_request = self.user_who_request(&target_channel);
-
                     if let Some(client_channel) =
                         self.chanmap.get_mut(&target_channel)
                     {
-                        client_channel.update_user_status(
-                            ok!(args.get(5)),
-                            ok!(args.get(6)),
-                            casemapping,
-                            isupport::get_bot_mode_char(&self.isupport),
-                        );
+                        let nick = ok!(args.get(5));
+                        let flags = ok!(args.get(6));
+                        let bot_mode_char =
+                            isupport::get_bot_mode_char(&self.isupport);
 
-                        if let Some(who_poll) = self
-                            .who_polls
-                            .iter_mut()
-                            .find(|who_poll| who_poll.channel == target_channel)
-                            && let WhoStatus::Requested(source, _, None) =
-                                &who_poll.status
+                        if self
+                            .capabilities
+                            .acknowledged(Capability::NoImplicitNames)
+                            && let Ok(mut user) = User::parse_from_whoreply(
+                                nick,
+                                flags,
+                                ok!(args.get(2)),
+                                ok!(args.get(3)),
+                                casemapping,
+                                isupport::get_prefix(&self.isupport),
+                            )
+                            && client_channel.users.resolve(&user).is_none()
                         {
-                            who_poll.status =
-                                WhoStatus::Receiving(source.clone(), None);
-                            log::trace!(
-                                "[{}] {channel} - WHO receiving...",
-                                self.server
+                            if flags.starts_with('G') {
+                                user.update_away(true);
+                            }
+                            if let Some(bot_char) = bot_mode_char {
+                                user.update_bot(flags.contains(bot_char));
+                            }
+                            client_channel.users.insert(user);
+                        } else {
+                            client_channel.update_user_status(
+                                nick,
+                                flags,
+                                casemapping,
+                                bot_mode_char,
                             );
                         }
+
+                        self.who_queue.receiving_channel_poll(
+                            &self.server,
+                            &target_channel,
+                        );
                     }
 
+                    let user_request = self.user_who_request(&target_channel);
                     if !user_request {
                         // User did not request, don't save to history
                         return Ok(vec![]);
@@ -1949,89 +1932,15 @@ impl Client {
 
                     if let Some(client_channel) =
                         self.chanmap.get_mut(&target_channel)
-                        && let Some(who_poll) = self
-                            .who_polls
-                            .iter_mut()
-                            .find(|who_poll| who_poll.channel == target_channel)
                     {
-                        match &who_poll.status {
-                            WhoStatus::Requested(
-                                source,
-                                _,
-                                Some(request_token),
-                            ) if matches!(source, WhoSource::Poll) => {
-                                if let Ok(token) =
-                                    ok!(args.get(1)).parse::<WhoToken>()
-                                    && *request_token == token
-                                {
-                                    who_poll.status = WhoStatus::Receiving(
-                                        source.clone(),
-                                        Some(*request_token),
-                                    );
-                                    log::trace!(
-                                        "[{}] {channel} - WHO receiving...",
-                                        self.server
-                                    );
-                                }
-                            }
-                            WhoStatus::Requested(
-                                WhoSource::User,
-                                _,
-                                Some(request_token),
-                            ) => {
-                                who_poll.status = WhoStatus::Receiving(
-                                    WhoSource::User,
-                                    Some(*request_token),
-                                );
-
-                                log::trace!(
-                                    "[{}] {channel} - WHO receiving...",
-                                    self.server
-                                );
-                            }
-                            _ => (),
-                        }
-
-                        if let WhoStatus::Receiving(WhoSource::Poll, Some(_)) =
-                            &who_poll.status
-                        {
-                            // Check token to ~ensure reply is to poll request
-                            if let Ok(token) =
-                                ok!(args.get(1)).parse::<WhoToken>()
-                            {
-                                if token == WhoXPollParameters::Default.token()
-                                {
-                                    client_channel.update_user_status(
-                                        ok!(args.get(3)),
-                                        ok!(args.get(4)),
-                                        casemapping,
-                                        isupport::get_bot_mode_char(
-                                            &self.isupport,
-                                        ),
-                                    );
-                                } else if token
-                                    == WhoXPollParameters::WithAccountName
-                                        .token()
-                                {
-                                    let user = ok!(args.get(3));
-
-                                    client_channel.update_user_status(
-                                        user,
-                                        ok!(args.get(4)),
-                                        casemapping,
-                                        isupport::get_bot_mode_char(
-                                            &self.isupport,
-                                        ),
-                                    );
-
-                                    client_channel.update_user_accountname(
-                                        user,
-                                        ok!(args.get(5)),
-                                        casemapping,
-                                    );
-                                }
-                            }
-                        }
+                        let _ = self.who_queue.handle_whox_reply(
+                            casemapping,
+                            &self.server,
+                            &self.isupport,
+                            &target_channel,
+                            client_channel,
+                            args,
+                        );
                     }
 
                     if !user_request {
@@ -2063,42 +1972,11 @@ impl Client {
                 ) {
                     let user_request = self.user_who_request(&target_channel);
 
-                    if let Some(pos) = self
-                        .who_polls
-                        .iter()
-                        .position(|who_poll| who_poll.channel == target_channel)
-                    {
-                        self.who_polls[pos].status = WhoStatus::Received;
-
-                        if let Some(who_poll) = self.who_polls.remove(pos)
-                            && self.chanmap.contains_key(&target_channel)
-                            && self.config.who_poll_enabled
-                        {
-                            self.who_polls.push_back(who_poll);
-                        }
-
-                        // Prioritize WHO requests due to joining a channel
-                        if let Some(pos) = self
-                            .who_polls
-                            .iter()
-                            .position(|who_poll| {
-                                matches!(who_poll.status, WhoStatus::Joined)
-                            })
-                            .or(self.who_polls.iter().position(|who_poll| {
-                                matches!(who_poll.status, WhoStatus::Received)
-                            }))
-                        {
-                            self.who_polls[pos].status =
-                                WhoStatus::Waiting(Instant::now());
-
-                            if pos != 0
-                                && let Some(who_poll) =
-                                    self.who_polls.remove(pos)
-                            {
-                                self.who_polls.push_front(who_poll);
-                            }
-                        }
-                    }
+                    self.who_queue.handle_end_who_reply(
+                        &self.chanmap,
+                        &self.config,
+                        &target_channel,
+                    );
 
                     log::trace!("[{}] {mask} - WHO done", self.server);
 
@@ -2109,7 +1987,7 @@ impl Client {
                     }
 
                     if !user_request {
-                        self.who_poll_interval.long_enough();
+                        self.who_queue.interval_long_enough();
 
                         // User did not request, don't save to history
                         return Ok(vec![]);
@@ -2135,24 +2013,10 @@ impl Client {
                         self.casemapping(),
                     );
 
-                    if let Some(pos) = self
-                        .who_polls
-                        .iter()
-                        .position(|who_poll| who_poll.channel == target_channel)
-                    {
-                        self.who_polls.remove(pos);
-                    } else {
-                        // User did not request, treat as part of rate-limiting response
-                        // (in conjunction with RPL_TRYAGAIN) and don't save to history.
-                        if let Some(who_poll) = self.who_polls.front_mut() {
-                            who_poll.status =
-                                WhoStatus::Waiting(Instant::now());
-                        }
-
-                        self.who_polls.iter_mut().skip(1).for_each(
-                            |who_poll| who_poll.status = WhoStatus::Received,
-                        );
-
+                    if !self.who_queue.remove_matching(|who_poll| {
+                        who_poll.channel == target_channel
+                    }) {
+                        self.who_queue.handle_who_rate_limit();
                         return Ok(vec![]);
                     }
                 }
@@ -2602,7 +2466,7 @@ impl Client {
 
                                                 self.anti_flood = None;
 
-                                                self.who_poll_interval.set_min(self.config.who_poll_interval);
+                                                self.who_queue.interval_set_min(self.config.who_poll_interval);
                                             }
                                             isupport::Parameter::BOUNCER_NETID(ref id) => {
                                                 match self.server.bouncer_netid() {
@@ -2911,15 +2775,15 @@ impl Client {
                 let command = ok!(args.get(1));
 
                 if command == "WHO" && self.config.who_poll_enabled {
-                    self.who_poll_interval.too_short();
+                    self.who_queue.interval_too_short();
 
                     log::debug!(
                         "[{}] WHO poll interval is too short → duration = {:?}",
                         self.server,
-                        self.who_poll_interval.duration()
+                        self.who_queue.interval_duration()
                     );
 
-                    if !self.who_polls.iter().any(|who_poll| {
+                    if !self.who_queue.any_matching(|who_poll| {
                         matches!(
                             who_poll.status,
                             WhoStatus::Requested(WhoSource::User, _, _)
@@ -3460,19 +3324,14 @@ impl Client {
     }
 
     fn user_who_request(&self, channel: &target::Channel) -> bool {
-        if let Some(who_poll) = self
-            .who_polls
-            .iter()
-            .find(|who_poll| who_poll.channel == *channel)
-        {
-            matches!(
-                who_poll.status,
-                WhoStatus::Requested(WhoSource::User, _, _)
-                    | WhoStatus::Receiving(WhoSource::User, _)
-            )
-        } else {
-            false
-        }
+        self.who_queue.any_matching(|who_poll| {
+            who_poll.channel == *channel
+                && matches!(
+                    who_poll.status,
+                    WhoStatus::Requested(WhoSource::User, _, _)
+                        | WhoStatus::Receiving(WhoSource::User, _)
+                )
+        })
     }
 
     pub fn chathistory_limit(&self) -> u16 {
@@ -4108,96 +3967,15 @@ impl Client {
             .for_each(|channel| prune_expired_typing(&mut channel.typing));
         prune_expired_querymap(&mut self.querymap);
 
-        if let Some(who_poll) = self.who_polls.front_mut() {
-            #[derive(Debug)]
-            enum Request {
-                Poll,
-                Retry,
-            }
-
-            let request = match &who_poll.status {
-                WhoStatus::Joined => {
-                    (self.capabilities.acknowledged(Capability::AwayNotify)
-                        || self.config.who_poll_enabled)
-                        .then_some(Request::Poll)
-                }
-                WhoStatus::Waiting(last) => {
-                    if self.capabilities.acknowledged(Capability::AwayNotify) {
-                        self.chanmap.get(&who_poll.channel).and_then(
-                            |channel| {
-                                (!channel.who_init
-                                    && (now.duration_since(*last)
-                                        >= self.who_poll_interval.duration()))
-                                .then_some(Request::Poll)
-                            },
-                        )
-                    } else {
-                        (self.config.who_poll_enabled
-                            && (now.duration_since(*last)
-                                >= self.who_poll_interval.duration()))
-                        .then_some(Request::Poll)
-                    }
-                }
-                WhoStatus::Requested(source, requested, _) => {
-                    if matches!(source, WhoSource::Poll)
-                        && !self.config.who_poll_enabled
-                    {
-                        None
-                    } else {
-                        (now.duration_since(*requested)
-                            >= 5 * self.who_poll_interval.duration())
-                        .then_some(Request::Retry)
-                    }
-                }
-                _ => None,
-            };
-
-            if let Some(request) = request {
-                log::trace!(
-                    "[{}] {} - WHO {}",
-                    self.server,
-                    who_poll.channel,
-                    match request {
-                        Request::Poll => "poll",
-                        Request::Retry => "retry",
-                    }
-                );
-
-                let message =
-                    if self.isupport.contains_key(&isupport::Kind::WHOX) {
-                        let whox_params = if self
-                            .capabilities
-                            .acknowledged(Capability::AccountNotify)
-                        {
-                            WhoXPollParameters::WithAccountName
-                        } else {
-                            WhoXPollParameters::Default
-                        };
-
-                        who_poll.status = WhoStatus::Requested(
-                            WhoSource::Poll,
-                            Instant::now(),
-                            Some(whox_params.token()),
-                        );
-
-                        command!(
-                            "WHO",
-                            who_poll.channel.to_string(),
-                            whox_params.fields().to_string(),
-                            whox_params.token().to_owned()
-                        )
-                    } else {
-                        who_poll.status = WhoStatus::Requested(
-                            WhoSource::Poll,
-                            Instant::now(),
-                            None,
-                        );
-
-                        command!("WHO", who_poll.channel.to_string())
-                    };
-
-                self.send(None, message.into(), TokenPriority::Low);
-            }
+        for (message, priority) in self.who_queue.tick(
+            &self.server,
+            &self.capabilities,
+            &self.chanmap,
+            &self.config,
+            &self.isupport,
+            now,
+        ) {
+            self.send(None, message, priority);
         }
 
         self.mode_requests.retain(|mode_request| {
@@ -4345,6 +4123,10 @@ impl Client {
 
     pub fn multiline_limits(&self) -> Option<MultilineLimits> {
         self.capabilities.multiline_limits()
+    }
+
+    pub fn prioritize_joined_who_poll(&mut self, channel: target::Channel) {
+        self.who_queue.prioritize_joined_who_poll(channel);
     }
 }
 
@@ -5439,27 +5221,6 @@ pub struct Topic {
     pub content: Option<message::Content>,
     pub who: Option<User>,
     pub time: Option<DateTime<Utc>>,
-}
-
-#[derive(Debug, Clone)]
-pub struct WhoPoll {
-    pub channel: target::Channel,
-    pub status: WhoStatus,
-}
-
-#[derive(Debug, Clone)]
-pub enum WhoStatus {
-    Requested(WhoSource, Instant, Option<WhoToken>),
-    Receiving(WhoSource, Option<WhoToken>),
-    Received,
-    Waiting(Instant),
-    Joined,
-}
-
-#[derive(Debug, Clone)]
-pub enum WhoSource {
-    User,
-    Poll,
 }
 
 #[derive(Debug, Clone)]