+++ data/src/isupport.rs
@@ -1132,6 +1132,7 @@ impl FromStr for WhoToken {
 pub enum WhoXPollParameters {
     Default,
     WithAccountName,
+    InitialJoin,
 }
 
 impl WhoXPollParameters {
@@ -1139,6 +1140,7 @@ impl WhoXPollParameters {
         match self {
             WhoXPollParameters::Default => "tcnf",
             WhoXPollParameters::WithAccountName => "tcnfa",
+            WhoXPollParameters::InitialJoin => "tcnfuha",
         }
     }
 
@@ -1149,6 +1151,9 @@ impl WhoXPollParameters {
             },
             WhoXPollParameters::WithAccountName => WhoToken {
                 digits: ['9', '9', '\0'],
+            },
+            WhoXPollParameters::InitialJoin => WhoToken {
+                digits: ['9', '9', '9'],
             },
         }
     }