+++ CHANGELOG.md
@@ -10,11 +10,13 @@ Added:
 - Support for IRCv3 `draft/message-redaction`
 - Setting to close queries when closing query buffer panes (`buffer.close.query`)
 - Support for IRCv3 metadata `display-name`, `avatar`, `pronouns`, `homepage`, `color`, `status` keys
+- IRCv3 `no-implicit-names` support
 
 Fixed:
 
 - `NICK` and `QUIT` messages returned by `draft/event-playback` (`chathistory`)
 - Backlog separator text colour now follows `buffer.backlog_rule`
+- Fix `tor` enabled builds
 - Distinguish between reacts in chathistory vs active history
 - Channel matching on single line only
 - Do not hide consecutive nicknames if the user's displayed access levels or bot mode changes